+++ templates/Beef.Template.Solution/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 Represents the **NuGet** versions.
 
+## v2.1.3
+- *Enhancement:* Added `PersonValidator` and `PersonTest` for end-to-end example.
+
 ## v2.1.2
 - *Enhancement:* Added new `--datasource` parameter. Choices are `Database` (default), `EntityFramework`, `Cosmos` or `None`.